+++ CHANGELOG.md
@@ -2,16 +2,19 @@
 
 ## [Unreleased]
 
-## [v1.0.0] - 2022-12-23
+## [v1.0.0] - 2023-01-02
 
 ### Added
 
+#### Function/ Method Arguments
+
 - Lint check that ensures all function/ method arguments are documented
-  (including `*args` and `**kwargs`)
+- Support `*args` and `**kwargs`
 - Support positional only arguments
 - Support keyword only arguments
 - Support ignoring `self` and `cls` arguments
 - Support for skipping test functions in test files
+- Support async functions
 
 [//]: # "Release links"
 [v1.0.0]: https://github.com/jdkandersson/flake8-docstrings-complete/releases/v1.0.0